services/streamlit/pyproject.toml
[project]
name = "dsdc"
version = "0.1.0"
description = "DataScientest Document Classification Project"
requires-python = ">=3.12"
readme = "README.md"
license = "MIT"
authors = [
  { name = "AL"},
  { name = "BH"},
]

dependencies = [
#    "clip",
#    "fastapi>=0.116.1",
    "matplotlib>=3.10.6",
#    "mlflow>=3.4.0",
#    "nltk>=3.9.1",
#    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "plotly>=6.3.0",
#    "psycopg2-binary>=2.9.10",
#    "pytesseract>=0.3.13",
#    "python-multipart>=0.0.20",
#    "pyyaml>=6.0.2",
    "requests>=2.32.5",
#    "scikit-learn>=1.7.2",
#    "sqlalchemy>=2.0.43",
    "streamlit>=1.50.0",
#    "tensorflow>=2.20.0",
#    "torch>=2.8.0",
#    "tqdm>=4.67.1",
#    "transformers>=4.56.1",
#    "uvicorn[standard]>=0.35.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["dsdc*"]

[dependency-groups]
dev = [
    "ipython>=9.5.0",
    "pytest>=8.4.2",
]
